Move module list to separate config file.
This patch moves the module list to a separate configuration file. This
is done to prevent evaluation of any PHP directives in the
`config/application.config.php` file when reading and then writing the
file on creation of a new API module and/or deletion of one. Since PHP
directives are evaluated, the evaluated values are then written to the
file, which may have unintended consequences, such as directory
expansion!